Move search methods to list VM
